Terminada la función generardatosbasicos

+++ src/modelo/Pelicula.java
@@ -7,7 +7,7 @@ public class Pelicula extends Contenido {
 	private String director;
 
 	public Pelicula(int idContenido, String titulo, Fecha fecha_lanzamiento,
-		int duracion, String genero, float imbdRating, float imbdVotes, int timesPlayed, int timesLiked, int timesNotLiked, float cristoflixRating, String director) {
+		int duracion, String genero, double imbdRating, double imbdVotes, int timesPlayed, int timesLiked, int timesNotLiked, double cristoflixRating, String director) {
 		super(idContenido, titulo, fecha_lanzamiento, duracion, genero, imbdRating, imbdVotes, timesPlayed, timesLiked, timesNotLiked, cristoflixRating); 
 		
 		this.director = director;
